Fix compiler errors
